# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Temple API",
    "description": "Docs for /contacts and /temples",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Endpoints for contact-related data"
    },
    {
      "name": "Temples",
      "description": "Endpoints for temple data"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/professional": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Return static professional contact info",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Return all contacts",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "favoriteColor": {
                  "type": "string",
                  "example": "Blue"
                },
                "birthday": {
                  "type": "string",
                  "example": "1990-01-01"
                }
              },
              "required": [
                "firstName",
                "lastName",
                "email",
                "favoriteColor",
                "birthday"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Return a single contact by MongoDB _id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "description": "Update an existing contact by _id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "favoriteColor": {
                  "type": "string",
                  "example": "Blue"
                },
                "birthday": {
                  "type": "string",
                  "example": "1990-01-01"
                }
              },
              "required": [
                "firstName",
                "lastName",
                "email",
                "favoriteColor",
                "birthday"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid ID or data"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "description": "Delete a contact by _id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/temples/": {
      "get": {
        "tags": [
          "Temples"
        ],
        "description": "Return all temples",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Temples"
        ],
        "description": "Create a new temple",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "temple_id": {
                  "type": "number",
                  "example": 9
                },
                "name": {
                  "type": "string",
                  "example": "New Temple"
                },
                "location": {
                  "type": "string",
                  "example": "Somewhere"
                },
                "dedicated": {
                  "type": "string",
                  "example": "TBD"
                },
                "additionalInfo": {
                  "type": "boolean",
                  "example": false
                }
              },
              "required": [
                "temple_id",
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/temples/{temple_id}": {
      "get": {
        "tags": [
          "Temples"
        ],
        "description": "Return a single temple by temple_id",
        "parameters": [
          {
            "name": "temple_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}